requirements.txt
torch
torchvision
pillow
numpy
opencv-python-headless
tensorboard
tqdm

# Required for .pkl to .bin
open3d